refactor: extract pause/resume cluster from state.py (#1271)
Pull the pause/resume pair (pause_game + resume_game) out of the
state.py God-Object into a new PauseResumeMixin (game/state_pause.py),
stacked on the round-lifecycle cut.

Pure code-move, API byte-identical: the methods carry over verbatim
(GamePhase imported lazily to avoid the state.py cycle, matching the
existing mixins). The _set_phase chokepoint, _timer_countdown,
_cancel_auto_advance, cancel_timer and end_round all stay on GameState
and are referenced via self. _previous_phase / pause_reason /
disconnected_admin_name stay as GameState instance attributes.

Mixin added alphabetically to the GameState base list + docstring;
the now-unused INTRO_DURATION_SECONDS import dropped from state.py.

state.py 1586 -> 1467 lines (-119).
